refactor: extract client_statistics.rs
