+++ components/restreamer/src/state.rs
@@ -1,10 +1,14 @@
 //! Application state.
 
+mod client_statistics;
 mod mixin;
 mod server_info;
 mod settings;
 
 pub use self::{
+    client_statistics::{
+        ClientStatistics, ClientStatisticsResponse, StatusStatistics,
+    },
     mixin::{Delay, Mixin, MixinId, MixinSrcUrl, Volume, VolumeLevel},
     server_info::ServerInfo,
     settings::Settings,
@@ -37,7 +41,6 @@ use url::Url;
 use uuid::Uuid;
 
 use crate::{display_panic, serde::is_false, spec, srs, Spec};
-use chrono::{DateTime, Utc};
 use std::collections::HashMap;
 
 /// Reactive application's state.
@@ -1954,65 +1957,4 @@ impl<'de> Deserialize<'de> for Label {
         Self::new(<Cow<'_, str>>::deserialize(deserializer)?)
             .ok_or_else(|| D::Error::custom("Not a valid Label"))
     }
-}
-
-/// Statistics of statuses in [`Input`]s or [`Output`]s of [`Client`]
-#[derive(Clone, Debug, Eq, GraphQLObject, PartialEq)]
-pub struct StatusStatistics {
-    /// Status of [`Input`]s or [`Output`]
-    pub status: Status,
-
-    /// Count of items having [`Status`]
-    /// GraphQLScalar requires i32 numbers
-    pub count: i32,
-}
-
-/// Information about status of all [`Input`]s and [`Output`]s and
-/// server health info (CPU usage, memory usage, etc.)
-#[derive(Clone, Debug, GraphQLObject, PartialEq)]
-pub struct ClientStatistics {
-    /// Client title
-    pub client_title: String,
-
-    /// Time when statistics was taken
-    pub timestamp: DateTime<Utc>,
-
-    /// Count of inputs grouped by status
-    pub inputs: Vec<StatusStatistics>,
-
-    /// Count of outputs grouped by status
-    pub outputs: Vec<StatusStatistics>,
-
-    /// Info about server info (CPU, Memory, Network)
-    pub server_info: ServerInfo,
-}
-
-impl ClientStatistics {
-    /// Creates a new [`ClientStatistics`] object with snapshot of
-    /// current client's statistics regarding [`Input`]s and [`Output`]s
-    #[must_use]
-    pub fn new(
-        client_title: String,
-        inputs: Vec<StatusStatistics>,
-        outputs: Vec<StatusStatistics>,
-        server_info: ServerInfo,
-    ) -> Self {
-        Self {
-            client_title,
-            timestamp: Utc::now(),
-            inputs,
-            outputs,
-            server_info,
-        }
-    }
-}
-
-/// Current state of [`ClientStatistics`] request
-#[derive(Clone, Debug, GraphQLObject, PartialEq)]
-pub struct ClientStatisticsResponse {
-    /// Statistics data
-    pub data: Option<ClientStatistics>,
-
-    /// The top-level errors returned by the server.
-    pub errors: Option<Vec<String>>,
 }
